libxtrans xwininfo libX11 libXau libXmu libXdmcp mesa mkfontscale xkeyboard-config git update 14 jan 2013
libxtrans: ec3136232f7ce930f9ca812b6ab42a71b60af4af
xwininfo: 3e60a26559221e561770710a8c4ed0b8ebc31afb
libX11: 3cd974b1d4d1fa6389d3695fa9fcc0c22a51d50c
libXau: 8570d287396934f26224c76d48d7f17d87380e72
libXmu: 9b253d99d5b4f3fbb681c2cb1b84f8f9acfee528
libXdmcp: ca65a92405500393f09d34388edbbf6350e6c146
mesa: e3e1ffb2520498584ef402213d0c8aa4303a46a3
mkfontscale: 1157b3039551b552b483f05f6a411e57941a87c0

+++ apps/xwininfo/xwininfo.c
@@ -1826,7 +1826,7 @@ get_net_wm_name (xcb_connection_t *gnwn_dpy, xcb_window_t win)
 #define UTF8_SHORT_TAIL 3
 #define UTF8_LONG_TAIL 4
 static int
-is_valid_utf8 (const char *string, int len)
+is_valid_utf8 (const char *string, size_t len)
 {
     unsigned long codepoint;
     int rem, i;
@@ -1947,7 +1947,7 @@ get_friendly_name (const char *string, const char *prefix)
     const char *name_start = string;
     char *lowered_name, *n;
     Bool first = True;
-    int prefix_len = strlen (prefix);
+    size_t prefix_len = strlen (prefix);
 
     if (strncmp (name_start, prefix, prefix_len) == 0) {
 	name_start += prefix_len;
